Separate deserialized_packet
