+++ core/src/lib.rs
@@ -22,6 +22,7 @@ pub mod commitment_service;
 pub mod completed_data_sets_service;
 pub mod consensus;
 pub mod cost_update_service;
+pub mod deserialized_packet;
 pub mod drop_bank_service;
 pub mod duplicate_repair_status;
 pub mod fetch_stage;
